Moved InventoryUpdate.java inside helper package
